[compiler]: Optimize parse error messages

+++ compiler/diagnostics/reporter.cpp
@@ -18,12 +18,13 @@ namespace swallow::compiler::diagnostics
       .with_type(ReportType::ERROR)
       .with_message(msg)
       .with_code(code)
-      .add_label(
-        LabelBuilder()
-          .with_message(labelMsg)
-          .with_span({ &Detail, loc.begin.column - 1, loc.end.column - 1 })
-          .with_color(ColorType::RED)
-          .build())
+      .add_label(LabelBuilder()
+                   .with_message(labelMsg)
+                   .with_span({ &Detail,
+                                static_cast<size_t>(loc.begin.column - 1),
+                                static_cast<size_t>(loc.end.column - 1) })
+                   .with_color(ColorType::RED)
+                   .build())
       .with_note(note)
       .build()
       .print(std::cout);
